wip: mount live preview to document root
