+++ packages/next/src/views/Document/getDocumentView.tsx
@@ -12,7 +12,6 @@ import type { ViewToRender } from './index.js'
 
 import { APIView as DefaultAPIView } from '../API/index.js'
 import { EditView as DefaultEditView } from '../Edit/index.js'
-import { LivePreviewView as DefaultLivePreviewView } from '../LivePreview/index.js'
 import { UnauthorizedViewWithGutter } from '../Unauthorized/index.js'
 import { VersionView as DefaultVersionView } from '../Version/index.js'
 import { VersionsView as DefaultVersionsView } from '../Versions/index.js'
@@ -125,17 +124,6 @@ export const getDocumentView = ({
             break
           }
 
-          case 'preview': {
-            // --> /collections/:collectionSlug/:id/preview
-            if (
-              (collectionConfig && collectionConfig?.admin?.livePreview) ||
-              config?.admin?.livePreview?.collections?.includes(collectionConfig?.slug)
-            ) {
-              View = getCustomViewByKey(views, 'livePreview') || DefaultLivePreviewView
-            }
-            break
-          }
-
           case 'versions': {
             // --> /collections/:collectionSlug/:id/versions
             if (docPermissions?.readVersions) {
@@ -242,18 +230,6 @@ export const getDocumentView = ({
           case 'api': {
             if (globalConfig?.admin?.hideAPIURL !== true) {
               View = getCustomViewByKey(views, 'api') || DefaultAPIView
-            }
-
-            break
-          }
-
-          case 'preview': {
-            // --> /globals/:globalSlug/preview
-            if (
-              (globalConfig && globalConfig?.admin?.livePreview) ||
-              config?.admin?.livePreview?.globals?.includes(globalConfig?.slug)
-            ) {
-              View = getCustomViewByKey(views, 'livePreview') || DefaultLivePreviewView
             }
 
             break
