+++ packages/ui/src/elements/LivePreview/Device/index.tsx
@@ -1,8 +1,8 @@
 'use client'
-import { useResize } from '@payloadcms/ui'
 import React, { useEffect } from 'react'
 
-import { useLivePreviewContext } from '../Context/context.js'
+import { useResize } from '../../../hooks/useResize.js'
+import { useLivePreviewContext } from '../../../providers/LivePreview/context.js'
 
 export const DeviceContainer: React.FC<{
   children: React.ReactNode
@@ -16,9 +16,9 @@ export const DeviceContainer: React.FC<{
 
   // Keep an accurate measurement of the actual device size as it is truly rendered
   // This is helpful when `sizes` are non-number units like percentages, etc.
-  // eslint-disable-next-line react-compiler/react-compiler -- TODO: fix
+
   const { size: measuredDeviceSize } = useResize(deviceFrameRef.current)
-  // eslint-disable-next-line react-compiler/react-compiler -- TODO: fix
+
   const { size: outerFrameSize } = useResize(outerFrameRef.current)
 
   let deviceIsLargerThanFrame: boolean = false
