+++ packages/ui/src/elements/LivePreview/DeviceContainer/index.tsx
@@ -1,7 +1,7 @@
 'use client'
 import React from 'react'
 
-import { useLivePreviewContext } from '../Context/context.js'
+import { useLivePreviewContext } from '../../../providers/LivePreview/context.js'
 
 export const DeviceContainer: React.FC<{
   children: React.ReactNode
